+++ chat/message/src/Message/Message.stories.tsx
@@ -5,7 +5,7 @@ import { WithMessageRating as MessageFeedbackStory } from '@lg-chat/message-feed
 import { storybookArgTypes, StoryMetaType } from '@lg-tools/storybook-utils';
 import { StoryFn } from '@storybook/react';
 
-import { Message, MessageSourceType } from '.';
+import { Message, MessageSourceType } from '..';
 
 const MarkdownText = `
 # Heading 1
