+++ chat/message/src/Message/Message.styles.ts
@@ -74,8 +74,3 @@ export const messageContainerWedgeStyles = {
     }
   `,
 };
-
-export const verifiedAnswerBannerStyles = css`
-  margin-top: -${spacing[200]}px;
-  margin-bottom: ${spacing[400]}px;
-`;
