+++ chat/message/src/MessageBanner/VerifiedAnswerBanner.styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@leafygreen-ui/emotion';
+import { spacing } from '@leafygreen-ui/tokens';
+
+export const verifiedAnswerBannerStyles = css`
+  margin-top: -${spacing[200]}px;
+  margin-bottom: ${spacing[400]}px;
+`;
