+++ chat/message/src/MessageBanner/index.ts
@@ -1,2 +1,3 @@
 export { MessageBanner } from './MessageBanner';
 export { type MessageBannerProps } from './MessageBanner.types';
+export { VerifiedAnswerBanner } from './VerifiedAnswerBanner';
